chore: refactor examples to be better isolated
