+++ examples/fastapi/pyproject.toml
@@ -1,0 +1,27 @@
+[tool.poetry]
+name = "fastapi-example"
+version = "0.1.0"
+description = "Example of using sundew to test applications using FastAPI"
+authors = ["devenjarvis <devenjarvis@gmail.com>"]
+readme = "README.md"
+packages = [{include = "fastapi_example"}]
+
+[tool.poetry.dependencies]
+python = "^3.9"
+fastapi = "^0.95.0"
+databases = "^0.7.0"
+uvicorn = "^0.21.1"
+aiosqlite = "^0.18.0"
+alembic = "^1.10.3"
+
+[tool.poetry.group.dev.dependencies]
+sundew = {path = "../../", develop = true}
+poethepoet = "*"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poe.tasks]
+run = "uvicorn app.main:app --reload"
+test = "sundew ./tests"
