+++ examples/simple_side_effects/tests/test_main.py
@@ -1,21 +1,22 @@
 import io
 
-from examples.examples import side_effects_example
+import main
+
 from sundew import test
 
-test(side_effects_example.concatenate)(
+test(main.concatenate)(
     kwargs={"a": "123", "b": "456"},
     returns="123456",
 )
-test(side_effects_example.reverse)(
+test(main.reverse)(
     kwargs={"a": "123"},
     returns="321",
 )
-test(side_effects_example.make_silly)(
+test(main.make_silly)(
     kwargs={"a": "123", "b": "456"},
     returns="321654",
 )
-test(side_effects_example.print_string)(
+test(main.print_string)(
     kwargs={"a": "123"},
     patches={"sys.stdout": io.StringIO()},
     side_effects=[
@@ -31,7 +32,7 @@ test(side_effects_example.print_string)(
     ],
 )
 
-test(side_effects_example.print_hello)(
+test(main.print_hello)(
     kwargs={"name": "Dave"},
     patches={"sys.stdout": io.StringIO()},
     side_effects=[
