+++ pyproject.toml
@@ -27,11 +27,7 @@ mypy = "*"
 ruff = "*"
 poethepoet = "*"
 types-cachetools = "*"
-fastapi = "^0.95.0"
-databases = "^0.7.0"
-uvicorn = "^0.21.1"
-aiosqlite = "^0.18.0"
-httpx = "^0.23.3"
+
 
 [build-system]
 requires = ["poetry-core"]
@@ -47,12 +43,21 @@ sundew = "sundew"
 test = "sundew ./tests"
 type = "mypy ."
 
+test_simple_async = "sundew ./examples/simple_async/tests"
+test_simple_example = "sundew ./examples/simple_example/tests"
+test_simple_side_effects = "sundew ./examples/simple_side_effects/tests"
+test_fastapi = "sundew ./examples/fastapi/tests"
+
+test_examples = ["test_simple_async", "test_simple_example", "test_simple_side_effects", "test_fastapi"]
+
+
 
 [tool.ruff]
 target-version = "py39"
 line-length = 100
 select = ["ALL"]
 ignore = ["D", "T", "ISC", "S101", "ANN101", "COM812", "TCH003", "PLR0913"]
+unfixable = ["ERA001"]
 
 [tool.mypy]
     [[tool.mypy.overrides]]
